newton_lab/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "newton-lab"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
  "hydra-core>=1.3",
  "numpy",
  "pyyaml",
  "torch",
  "warp-lang",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["newton_lab*"]
